+++ src/ManagedModels/States/Archivable/ArchiveScope.php
@@ -13,6 +13,6 @@ class ArchiveScope implements Scope
 {
     public function apply(Builder $builder, Model $model)
     {
-        $builder->where(PageState::KEY, '<>', PageState::archived);
+        $builder->where($model->getTable() .'.'. PageState::KEY, '<>', PageState::archived);
     }
 }