+++ src/ManagedModels/States/SimpleState/UsesSimpleState.php
@@ -55,7 +55,7 @@ trait UsesSimpleState
             return;
         }
 
-        $query->whereIn($this->getStateAttribute(), $this->onlineStates());
+        $query->whereIn($this->getTable().'.'.$this->getStateAttribute(), $this->onlineStates());
     }
 
     protected function getStateAttribute(): string